% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_regression_fn.R
\name{bizicount}
\alias{bizicount}
\title{Maximum likelihood estimation of copula-based bivariate zero-inflated
(and non-inflated) count models}
\usage{
bizicount(
  fmla1,
  fmla2,
  data,
  cop = "gaus",
  margins = c("pois", "pois"),
  link.ct = c("log", "log"),
  link.zi = c("logit", "logit"),
  starts = NULL,
  keep = F,
  subset,
  na.action,
  weights,
  frech.min = 1e-07,
  pmf.min = 1e-07,
  ...
)
}
\arguments{
\item{fmla1}{A \code{formula} for the first margin. If non-inflated, of the form
\code{y ~ x1 + x2}; if inflated, of the form \code{y ~ x1 + x2 | z1 + z2}, where \code{y}
is the outcome for the first margin, \code{x} are covariates for count
parameters, and \code{z} are covariates for zero-inflated parameters.}

\item{fmla2}{A \code{formula} for the second margin. See \code{fmla1}.}

\item{data}{A dataframe containing the response variables, covariates, and
offsets for the model. If \code{NULL}, these quantities are searched for in the
parent environment.}

\item{cop}{Character string specifying the copula to be used. One of \code{"gaus"}
or \code{"frank."} Partial matching supported.}

\item{margins}{Length 2 character vector specifying the marginal
distributions for each outcome. Each of the two elements must be one of
\code{c("pois", "nbinom", "zip", "zinb")}, and must be consistent with its
corresponding formula (i.e., zero-inflated margins with zero-inflated
formulas).}

\item{link.ct}{Length 2 character string specifying the link function used
for the count portion of each margin (if zero-inflated, otherwise it is the
link function for the conditional mean). One of \code{c("log", "identity", "sqrt")}.}

\item{link.zi}{Length 2 character string specifying the link function used
for the zero-inflation portion of each margin. One of \code{c("logit", "probit", "cauchit", "log", "cloglog")}. Ignored if corresponding \code{margins} entry is
not zero-inflated.}

\item{starts}{Numeric vector of starting values for parameter estimates. See
details on the correct order for the values in this vector. If \code{NULL},
starting values are obtained automatically by a univariate regression fit.}

\item{keep}{Logical indicating whether to keep the model matrix in the
returned model object. Defaults to \code{FALSE} to conserve memory. NOTE: This
must be set to \code{TRUE} to use \code{\link[texreg]{texreg}} or \code{\link{make.DHARMa}} functions with
\code{bizicount} objects.}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process.}

\item{na.action}{a function which indicates what should happen when the data
contain NAs. The default is set by the na.action setting of options, and is
na.fail if that is unset. The ‘factory-fresh’ default is na.omit. Another
possible value is NULL, no action. Value na.exclude can be useful.}

\item{weights}{An optional numeric vector of weights for each observation.}

\item{frech.min}{Lower boundary for Frechet-Hoeffding bounds on copula CDF.
Used for computational purposes to prevent over/underflow in likelihood
search. See details.}

\item{pmf.min}{Lower boundary on copula PMF evaluations. Used for
computational purposes to prevent over/underflow in likelihood search. See
details.}

\item{...}{Additional arguments to be passed on to the quasi-newton fitting
function, \code{\link[stats]{nlm}}}
}
\description{
The main bivariate regression function of the 'bizicount'
package. Estimates copula-based bivariate zero-inflated (and non-inflated)
count models via maximum likelihood. Supports the Frank and Gaussian
copulas, zero-inflated Poisson and negative binomial margins (and their
non-inflated counterparts).
}
\details{
These are details.
}
