% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bizicount_methods.R
\name{extract.bizicount}
\alias{extract.bizicount}
\title{Texreg for bizicount objects}
\usage{
\method{extract}{bizicount}(model, CI = NULL, id = T)
}
\arguments{
\item{model}{A \code{\link{bizicount-class}} model object (S3).}

\item{CI}{The two-tailed confidence level, if confidence intervals are
desired in the texreg object, otherwise \code{NULL}.}

\item{id}{Logical indicating whether to prepend equation identifiers to
coefficient names (\code{ct_} for count parameters, \code{zi_} for zero-inflated parameters)}
}
\value{
A \code{link[texreg]{texreg-class}} object, as produced by
\code{link[texreg]{createTexreg}}, which can interface with all of that
package's methods.
}
\description{
This is a method for the \code{\link[texreg]{extract}} generic
to be used with \code{\link{bizicount-class}} objects that are output from the
\code{\link{bizicount}} function.
}
\seealso{
\code{\link[texreg]{extract}}, \code{\link[texreg]{createTexreg}},
\code{\link[bizicount]{bizicount}}
}
\author{
John Niehaus
}
