% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bizicount_methods.R
\name{make_DHARMa}
\alias{make_DHARMa}
\title{DHARMa-class objects from bizicount models}
\usage{
make_DHARMa(object, nsim = 250, seed = 123, method = "PIT")
}
\arguments{
\item{object}{A \code{\link{bizicount-class}} object, as returned by \link{bizicount}.}

\item{nsim}{Number of simulated responses from the fitted model to use for diagnostics.}

\item{seed}{Random seed for simulating from fitted model.}

\item{method}{See \code{\link[DHARMa]{createDHARMa}}.}
}
\value{
A list of \code{\link[DHARMa]{DHARMa}} objects.
}
\description{
A wrapper around the \code{\link[DHARMa]{DHARMa}} package's
\code{\link[DHARMa]{createDHARMa}} function. Creates a list of DHARMa
objects, one for each margin of a \code{\link{bizicount-class}} object, using
simulated responses from \code{\link{simulate.bizicount}}.
}
\note{
This is merely a wrapper around the \code{\link[DHARMa]{createDHARMa}}
function to conveniently get DHARMa objects for each margin of a bizicount
model.
}
\seealso{
\code{\link[DHARMa]{DHARMa}}, \code{\link[DHARMa]{createDHARMa}},
\code{\link{simulate.bizicount}}
}
\author{
John Niehaus
}
