% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zicreg-methods.R
\name{predict.zicreg}
\alias{predict.zicreg}
\title{Predictions for univariate zero-inflated count regression models}
\usage{
\method{predict}{zicreg}(object, newdata = NULL, y.new = NULL, type = "mean", ...)
}
\arguments{
\item{object}{A fitted \code{\link{zic.reg}} object.}

\item{newdata}{A \code{\link[base]{data.frame}} containing new values of the same covariates appearing in fitted model.}

\item{y.new}{An optional vector of new response values, used only for \code{type = "prob"}.}

\item{type}{String, one of \code{c("mean", "prob", "psi", "lambda")}. \code{"mean"}
will predict the conditional mean of the mixture distribution, \code{"prob"}
will predict the probability of a new response value, \code{"psi"} will predict
the probability of zero-inflation, and \code{"lambda"} will predict the mean of
the count portion of the mixture distribution. NOTE: Setting \code{type = "mean"} and leaving \code{newdata = NULL} is the same as calling
\code{fitted(object)}.}

\item{...}{Ignored.}
}
\description{
Predicts the mean, probability, count mean, or zero-inflation
probability for new data using parameters from a fitted zero-inflated count
regression model.
}
