% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bizicount_methods.R
\name{simulate.bizicount}
\alias{simulate.bizicount}
\title{Simulating response values using parameters from fitted bizicount models}
\usage{
\method{simulate}{bizicount}(object, nsim = 250, seed = 123, ...)
}
\arguments{
\item{object}{A fitted \code{\link{bizicount-class}} object, as returned by
\code{\link{bizicount}}.}

\item{nsim}{Number of simulated response values from the fitted model. E.g.,
\code{nsim = 250} will simulate each observation 250 times, for \eqn{n \times
  250} total observations.}

\item{seed}{Seed used for simulating from fitted model. If \code{NULL}, no seed is
set.}

\item{...}{Ignored.}
}
\value{
A length 2 list, with each entry containing a numeric \eqn{n \times
  nsim} matrix for each margin of the bizicount model. Rows index
the observation, and columns index the simulated dataset number.
}
\description{
Simulates random response values using the fitted conditional
mean function for each margin of a \code{\link{bizicount-class}} object.
Primarily for use with the \code{\link[DHARMa]{DHARMa}} package.
}
\author{
John Niehaus
}
