% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zicreg_methods.R
\name{simulate.zicreg}
\alias{simulate.zicreg}
\title{Simulating response values from fitted univariate zero-inflated count
regression model}
\usage{
\method{simulate}{zicreg}(object, nsim = 250, seed = 123, ...)
}
\arguments{
\item{object}{A \code{\link{zicreg-class}} omodel bject, as returned by \code{\link{zic.reg}}.}

\item{nsim}{Number of simulated datasets to create.}

\item{seed}{Random seed for random number generation in simulations. If
\code{NULL}, no seed is set.}

\item{...}{Ignored.}
}
\value{
A numeric \eqn{n \times nsim} matrix, with rows indexing
observations, and columns indexing the simulation number.
}
\description{
Simulates responses using the fitted parameters from a
\code{\link{zicreg-class}} object, as returned by \code{\link{zic.reg}}.
Primarily useful for methods found in \code{\link[DHARMa]{DHARMa}} package.
}
\author{
John Niehaus
}
