% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zicreg.R
\name{zic.reg}
\alias{zic.reg}
\title{Univariate zero-inflated Poisson and negative binomial regression models}
\usage{
zic.reg(
  fmla = NULL,
  data,
  dist = "pois",
  link.ct = "log",
  link.zi = "logit",
  optimizer = "nlm",
  starts = NULL,
  subset,
  na.action,
  weights = rep(1, length(y)),
  X = NULL,
  z = NULL,
  y = NULL,
  offset.ct = NULL,
  offset.zi = NULL,
  warn.parent = T,
  keep = F,
  ...
)
}
\arguments{
\item{fmla}{A \code{\link[stats]{formula}} of the form \verb{y ~ x_1 + x_2 + ... + x_n + offset(count_var) | z_1  + ... z_n + offset(zi_var)}, where the \code{x}
values are covariates in the count portion of the model, and \code{z} are in the
zero-inflation portion. The \code{z} and \code{x} variables can be the same. If \code{NULL},
design matrices, the response vector, and offsets can be entered directly; see
\code{X}, \code{z}, \code{y}, \code{offset.ct}, and \code{offset.zi} below.}

\item{data}{A \code{\link[base]{data.frame}} containing all variables
appearing in \code{fmla}, including offsets. If not specified, variables are
searched for in parent environment.}

\item{dist}{The distribution used for the count portion of the zero-inflated
mixture. One of \code{c("pois", "nbinom")}, partial matching supported.}

\item{link.ct}{String specifying the link function used for the count portion
of the mixture distribution. One of \code{c("log", "identity", "sqrt")}.
See \code{\link[stats]{family}}.}

\item{link.zi}{Character string specifying the link function used for the
zero-inflation portion of the mixture distribution. One of \code{c("logit", "probit", "cauchit", "log", "cloglog")}. See \code{\link[stats]{family}}.}

\item{optimizer}{String specifying the optimizer to be used for fitting, one
of \code{c("nlm", "optim")}. If \code{"optim"}, defaults to \code{method="BFGS"}.}

\item{starts}{Optional vector of starting values used for the numerical
optimization procedure. Should have count parameters first (with intercept
first, if applicable), followed by zero-inflated parameters (with intercept
first, if applicable), and the inverse dispersion parameter last (if
applicable).}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process.}

\item{na.action}{A function which indicates what should happen when the data
contain NAs. The default is set by the na.action setting of options, and is
na.fail if that is unset. The ‘factory-fresh’ default is na.omit. Another
possible value is NULL, no action. Value na.exclude can be useful.}

\item{weights}{An optional numeric vector of weights for each observation.}

\item{X, z}{If \code{fmla = NULL}, these are the design matrices of covariates for
the count and zero-inflation portions, respectively. Both require no
missingness. Similar in spirit to \code{\link[stats]{glm.fit}} in that it
can be faster for larger datasets because it bypasses model matrix
creation.}

\item{y}{If \code{fmla = NULL}, a vector containing the response variable.}

\item{offset.ct, offset.zi}{If \code{fmla = NULL}, vectors containing the
(constant) offset for the count and zero-inflated portions, respectively.
Must be equal in length to \code{y}, and row-dim of \code{X}, \code{z}. If left \code{NULL},
defaults to \code{rep(0, length(y))}.}

\item{warn.parent}{Logical indicating whether to warn about \code{data} not
being supplied.}

\item{keep}{Logical indicating whether to keep the model matrices in
the returned model object. Must be \code{TRUE} to use
\code{\link[=DHARMa]{DHARMa}} and \code{\link[texreg]{texreg}} with the
model object, e.g., via \code{\link{simulate.zicreg}} and
\code{\link{extract.zicreg}}.}

\item{...}{Additional arguments to pass on to the chosen optimizer, either
\code{\link[stats]{nlm}} or \code{\link[stats]{optim}}. See 'Examples'.}
}
\value{
An S3 object (and list) of class \code{zicreg}.
}
\description{
This function from the \code{\link{bizicount}} package estimates
univariate zero-inflated Poisson and negative binomial regression models
via maximum likelihood using either the \code{\link[stats]{nlm}} or
\code{\link[stats]{optim}} optimization functions.  It's class has
associated \code{\link[stats]{simulate}} methods for post-estimation
diagnostics using the \code{\link[=DHARMa]{DHARMa}} package, as well as an
\code{\link[texreg]{extract}} method for printing professional tables using
\code{\link[texreg]{texreg}}. Visit the 'See Also' section for links to these
methods for \code{zicreg} objects.
}
\examples{
#zic.reg examples

## ZIP example
# Simulate some zip data
n=1000
x = cbind(1, rnorm(n))
z = cbind(1, rbeta(n, 4, 8))
b = c(1, 2.2)
g = c(-1, 1.7)
lam = exp(x \%*\% b)
psi = plogis(z \%*\% g)

y = bizicount::rzip(n, lambda = lam, psi=psi)


# estimate zip model using NLM, no data.frame

mod = zic.reg(y ~ x[,-1] | z[,-1])


# estimate zip using NLM, adjust stepmax via ... param

mod = zic.reg(y ~ x[,-1] | z[,-1], stepmax = .5)


# estimate zip using optim

mod = zic.reg(y ~ x[,-1] | z[,-1], optimizer = "optim")


# pass different method, reltol to optim using ... param

mod = zic.reg(y ~ x[,-1] | z[,-1],
        optimizer = "optim",
        method = "Nelder-Mead",
        control = list(reltol = 1e-10)
        )

# No formula, specify design matrices and offsets.
zic.reg(y=y, X=x, z=z)



## ZINB example
# simulate zinb data

disp = .5

y = bizicount::rzinb(n, psi = psi, size = disp, mu=lam)


# zinb model, use keep = TRUE for post-estimation methods

mod = zic.reg(y ~ x[,-1] | z[,-1], dist = "n", keep = TRUE)



## Make DHARMa object for diagnostics

# simulate from fitted model

sims = simulate(mod)

# Make dharma object

dharm = DHARMa::createDHARMa(
     simulatedResponse = sims,
     observedResponse = y,
     fittedPredictedResponse = fitted(mod),
     integerResponse = TRUE,
     method = "PIT"
)

# Plot the DHARMa object, shows that model fit is poor
plot(dharm)
DHARMa::testResiduals(dharm)



## Output to table with texreg

# extract information

tr_obj_se = extract(mod)
tr_obj_ci = extract(mod, CI = .95)

# output to latex, single table

texreg(list(tr_obj_se, tr_obj_ci))

# output to plain text, multiple tables

screenreg(tr_obj_se)
screenreg(tr_obj_ci)

}
\references{
Lambert, Diane. "Zero-inflated Poisson regression, with an
application to defects in manufacturing." Technometrics 34.1 (1992): 1-14.
}
\seealso{
\code{\link{simulate.zicreg}}, \code{\link{extract.zicreg}}.
}
\author{
John Niehaus
}
