% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bizicount_methods.R
\name{extract.bizicount}
\alias{extract.bizicount}
\title{Texreg for bizicount objects}
\usage{
\method{extract}{bizicount}(model, CI = NULL, id = T)
}
\arguments{
\item{model}{A \code{\link{bizicount-class}} model object (S3).}

\item{CI}{The two-tailed confidence level, if confidence intervals are
desired in the texreg object, otherwise \code{NULL}.}

\item{id}{Logical indicating whether to prepend equation identifiers to
coefficient names (\code{ct_} for count parameters, \code{zi_} for zero-inflated parameters)}
}
\value{
A \code{link[texreg]{texreg-class}} object, as produced by
\code{link[texreg]{createTexreg}}, which can interface with all of that
package's methods.
}
\description{
This is a method for the \code{\link[texreg]{extract}} generic
to be used with \code{\link{bizicount-class}} objects that are output from the
\code{\link{bizicount}} function.
}
\examples{
# bizicount example
set.seed(123)
n = 500

# define a function to simulate from a gaussian copula
# first margin is zero-inflated negative binomial (zinb)
# second margin is zero-inflated poisson (zip)
# Note: marginal distributions are hard-coded in function, including
# inverse dispersion parameter for zinb.
gen = function(n,
               b1,
               b2,
               g1,
               g2,
               dep) {

     k1 = length(b1)
     k2 = length(b2)

     X1 = cbind(1, matrix(rbinom(n * (k1 - 1), 1, .5), ncol = k1 - 1))
     X2 = cbind(1, matrix(rexp(n * (k2 - 1), 3), ncol = k2 - 1))

     lam1 = exp(X1 \%*\% b1)
     lam2 = exp(X2 \%*\% b2)

     Z1 = cbind(1, matrix(runif(n * (k1 - 1), -1, 1), ncol = k1 - 1))
     Z2 = cbind(1, matrix(rnorm(n * (k2 - 1)), ncol = k2 - 1))

     psi1 = plogis(Z1 \%*\% g1)
     psi2 = plogis(Z2 \%*\% g2)

     norm_vars = MASS::mvrnorm(
          n,
          mu = c(0, 0),
          Sigma = matrix(c(1, dep, dep, 1), ncol =2)
          )

     U = pnorm(norm_vars)

     y1 =  qzinb(U[, 1],
                 mu = lam1,
                 psi = psi1,
                 size = .3)
     y2 =  qzip(U[, 2],
                lambda = lam2,
                psi = psi2)

     dat = data.frame(
          X1 = X1[, -1],
          X2 = X2[, -1],
          Z1 = Z1[, -1],
          Z2 = Z2[, -1],
          y1,
          y2,
          lam1,
          lam2,
          psi1,
          psi2
     )
     return(dat)
}

# define parameters
b1 = c(1, -2, 3)
b2 = c(-1, 3, 1)
g1 = c(2, -1.5, 2)
g2 = c(-1, -3.75, 1.25)
rho = .5

# generate data
dat = gen(n, b1, b2, g1, g2, rho)
f1 = y1 ~ X1.1 + X1.2 | Z1.1 + Z1.2
f2 = y2 ~ X2.1 + X2.2 | Z2.1 + Z2.2

# estimate model
mod = bizicount(f1, f2, dat, cop = "g", margins = c("zinb", "zip"), keep=TRUE)

# diagnose model with DHARMa (will produce 6 plots, and print to console)
# see end for simulate.bizicount example.

dharm = make_DHARMa(mod)
lapply(dharm, plot)
lapply(dharm, DHARMa::testResiduals)

# extract texreg objects, one with SEs, one with CIs
tr_obj_se = extract(mod)
tr_obj_ci = extract(mod, CI = .95)

# output to latex, single table.
# Note use of c(), because tr_obj_se, tr_obj_ci are lists.
texreg(c(tr_obj_se, tr_obj_ci))

# output as plaintext, two tables
screenreg(tr_obj_se)
screenreg(tr_obj_ci)


}
\seealso{
\code{\link[texreg]{extract}}, \code{\link[texreg]{createTexreg}},
\code{\link[bizicount]{bizicount}}
}
\author{
John Niehaus
}
